grammar now parses input as binary-ish + ~matches the Fluka Manual
This  grammar seems to capture every syntactically correct text I can
throw at it.

The fundamental component is still the unary expression, but now
binary expressions are expressed as a combinations of a unary
expression and a binary expression.  This use of unaryExpression as
the atom is necessary to resolve ambiguous  expressions such as
"-A -B", which have  no meaning in its own right as a solid, and is
generally part of an effort to stitch GDML together with fluka.
Additionally, it matches Fluka, which itself consists of unary
expressions as the atomic expression.