refactored so that Body.resize mutates in place + code in Zone
code in zone to support this new method of recursively scaling zones.
This is better as it is simpler to implement, and the bodies are
mutated in a way that doesn't actually make them need to be
reinstantiated..  all information is kept, just transient hidden
attributes are changed.  this is the simplest way forwards without a
big change.

Note:  still doesn't actually work for subzones yet.