Fixed bug with malformation of RPP minimisation
Malformation in the sense that RPPs would become progressively smaller
if the upper bound was negative or the lower bound was
positive (whislt being larger than the extent of the final solid).
This is now fixed--the RPP minimisations converges immediately to a
fraction larger than the extent of the resulting mesh.