Fixed bug where RPP minimisation was polluted across >1 zones
In one zone it may have been appropriate to minimise an RPP by X
amount.  But then in a totally separate zone, where the RPP perhaps
shouldn't be minimised at all, it would nonetheless remain minimised.
This would lead to null meshes, and is... bad.

This is a temporary fix.  Really this needs refactoring, I don't like
this pattern of calling _apply_crude_scale on finite bodies.  It is
stupid.  Oh well.