regions dict to collections.OrderedDict in FlukaRegionVisitor
This is necessary because in Fluka, behind the scenes, the order of the regions
does matter, as each region is assigned an index.  This index may then be used
to assign, for example, materials to a sequence of regions, defined by lower and
upper bounds.  Therefore the order must be preserved when reading the regions in.