error listener to make the lexer/parser throw exceptions
By default the ANTLR4 parser and/or lexer will try and recover, often
resulting in a rubbish/nonsense AST, causing errors far removed from
the original cause.  I don't want this.  I want it to fail as noisily
as possible.  The class "SensitiveErrorListener" will enable the lexer
and parser to throw exceptions when it fails to lex / parse.