by default no length safety when doing boolean ops
explicitly length safety when getting overlaps w/ get_overlap

explicitly length safety when doing _get_overlap

which is separate for reasons unknown

only apply length safety to a union if optimise=True

this is the "rule"..  I think I stick to this.

more explicit safety checking

previously (I THINK) relied on default kwargs.  hopefully this is the
same but clearer.