Added connected_zones method + a test + removed redundant method
connected_zones returns a generator of sets of zones which are
overlapping as part of a union.  Two zones do not need to be directly
overlapping with another so long as, for example, they are mutually
overlapping with a third solid.  disjoint_zones has been removed as
connected_zones is many times more useful.