check_overlaps optimisation based on caching + bounding boxes.
Optimised booleans are cached.  This saves calculating the optimised
boolean for every individual region.  This saves a lot of time.

Along side the the optimised booleans are the extents.  Caching the
extents allows one to perform a basic overlap check just by comparing
the dimensions of these boxes, potentially saving many costly intersections.