Fixed bug where underestimating the size of a PLA.
Basically this was because I was using an extent (basically a box
without any rotation) to determine the size of a box which may well
have rotation.  So if the box was rotated then it would be too small,
leading to malformed geometry.  Introducint a factor of sqrt(3)
ensures that regardless of orientation, the box will contain the extent.