code for building up a map of the minimum extent for each body.
Build a map of bodies to regions
Build a map of regions to extents
Combine both maps to build a map of bodies to region extents.

Loop over bodies to region extents, combine extents into the minimum
extent which encompasses all the extents.

TODO: This can then be referred to when using the body an calling
e.g. geant4Solid on that body to make the body as small as possible,
whilst being big enough for every single region in which it is used.