fix bug in MultiUnion.mesh method
It was implemented like a conventional Boolean solid, the first solids has
no rotation and translation, with all subsequent solids rotated/translated
with respect to that.  But this doesn't match geant4's behaviour, there is
no priveleged solid, they all have a rotation and translation.  This is
further supported by looking in G4MultiUnion.hh, there is simply "AddNode"
for adding a solid with a rotation/translation, and there is no ctor with a
first solid.  So this mesh implementation is now the correct one.