my attempt at a slightly better introduction to gdml2fluka section.
describe what a region is as soon as we refer to it.  (slight reordering)

rewording of the description of a region so that we mention zones, which I
think are important, and a more top-down description, starting with the
definition of a region, then to a zone, and then explaining the zone in
terms of primtives.