Default name multi union node pos & rot if not provided in GDML
Scheme will complain if the positions and rotations don't have names.  We
could just make sure to name them elsewhere, but the better policy is
probably to simply ensure that we never emit non-conforming GDML.  This is
because we already provide automatic naming for each node, so we may as
well provide automatic naming for each node pos&rot based on that name.